Consolidate EditSayings VC and update experimental empty states to include actions
