+++ Vocable/Common/Views/EmptyStateView.swift
@@ -8,7 +8,27 @@
 
 import UIKit
 
-final class EmptyStateView: UIView {
+
+private class EmptyStateButton: GazeableButton {
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        commonInit()
+    }
+
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        commonInit()
+    }
+
+    private func commonInit() {
+        contentEdgeInsets = .init(top: 16, left: 16, bottom: 16, right: 16)
+    }
+}
+
+class EmptyStateView: UIView {
+
+    typealias ButtonConfiguration = (title: String, action: () -> Void)
 
     var image: UIImage? {
         get {
@@ -46,17 +66,21 @@ final class EmptyStateView: UIView {
 
     private let imageView = UIImageView(frame: .zero)
     private let label = UILabel(frame: .zero)
+    private let button = EmptyStateButton(frame: .zero)
+    private let action: ButtonConfiguration?
 
     private lazy var stackView = UIStackView(arrangedSubviews: [self.imageView, self.label])
 
-    init(text: String, image: UIImage? = nil) {
+    init(text: String, image: UIImage? = nil, action: ButtonConfiguration? = nil) {
+        self.action = action
         super.init(frame: .zero)
         self.text = text
         self.image = image
         commonInit()
     }
 
-    init(attributedText: NSAttributedString) {
+    init(attributedText: NSAttributedString, action: ButtonConfiguration? = nil) {
+        self.action = action
         super.init(frame: .zero)
         self.text = nil
         self.image = nil
@@ -65,6 +89,7 @@ final class EmptyStateView: UIView {
     }
 
     required init?(coder: NSCoder) {
+        self.action = nil
         super.init(coder: coder)
         commonInit()
     }
@@ -78,6 +103,17 @@ final class EmptyStateView: UIView {
         stackView.axis = .vertical
         stackView.alignment = .center
 
+        if let action = action {
+            let font = UIFont.boldSystemFont(ofSize: 18)
+            let attributed = NSAttributedString(string: action.title,
+                                                attributes: [.font: font, .foregroundColor: UIColor.defaultTextColor])
+            button.setAttributedTitle(attributed, for: .normal)
+            button.addTarget(self,
+                             action: #selector(handleButton(_:)),
+                             for: .primaryActionTriggered)
+            stackView.addArrangedSubview(button)
+        }
+
         stackView.translatesAutoresizingMaskIntoConstraints = false
         addSubview(stackView)
         NSLayoutConstraint.activate([
@@ -89,15 +125,35 @@ final class EmptyStateView: UIView {
             stackView.centerXAnchor.constraint(equalTo: centerXAnchor)
         ])
 
-        imageView.tintColor = .defaultTextColor
+        let color = UIColor.defaultTextColor
+        imageView.tintColor = color
 
-        label.textColor = .defaultTextColor
-        label.font = .boldSystemFont(ofSize: 28)
+        label.textColor = color
         label.textAlignment = .center
         label.numberOfLines = 0
 
         NSLayoutConstraint.activate([
             label.widthAnchor.constraint(equalTo: readableContentGuide.widthAnchor)
         ])
+
+        updateContentForCurrentTraitCollection()
+    }
+
+    override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
+        super.traitCollectionDidChange(previousTraitCollection)
+        updateContentForCurrentTraitCollection()
+    }
+
+    private func updateContentForCurrentTraitCollection() {
+
+        if [traitCollection.horizontalSizeClass, traitCollection.verticalSizeClass].contains(.compact) {
+            label.font = .boldSystemFont(ofSize: 28)
+        } else {
+            label.font = .boldSystemFont(ofSize: 20)
+        }
+    }
+
+    @objc private func handleButton(_ sender: GazeableButton) {
+        action?.action()
     }
 }
