+++ Vocable/Features/Presets/Pagination/PresetCollectionViewController.swift
@@ -92,7 +92,7 @@ class PresetCollectionViewController: CarouselGridCollectionViewController, NSFe
         }
         updateDataSource(animated: true)
 
-        self.scrollToMiddleSection()
+//        self.scrollToMiddleSection()
     }
     
     override func viewWillAppear(_ animated: Bool) {
@@ -169,6 +169,48 @@ class PresetCollectionViewController: CarouselGridCollectionViewController, NSFe
         dataSourceProxy.apply(snapshot,
                               animatingDifferences: animated,
                               completion: completion)
+
+        let category = Category.fetch(.userFavorites, in: fetchedResultsController!.managedObjectContext)
+        if content.isEmpty, category.objectID == categoryID {
+            installEmptyStateIfNeeded()
+        } else {
+            removeEmptyStateIfNeeded()
+        }
+    }
+
+    private func addPhrasePressed() {
+        let storyboard = UIStoryboard(name: "EditTextViewController", bundle: nil)
+        let vc = storyboard.instantiateViewController(identifier: "EditTextViewController") as! EditTextViewController
+        vc.editTextCompletionHandler = { (newText) -> Void in
+            let context = NSPersistentContainer.shared.viewContext
+
+            _ = Phrase.create(withUserEntry: newText, in: context)
+            do {
+                try context.save()
+
+                let alertMessage: String = {
+                    let format = NSLocalizedString("phrase_editor.toast.successfully_saved_to_favorites.title_format", comment: "Saved to user favorites category toast title")
+                    let categoryName = Category.userFavoritesCategoryName()
+                    return String.localizedStringWithFormat(format, categoryName)
+                }()
+
+                ToastWindow.shared.presentEphemeralToast(withTitle: alertMessage)
+            } catch {
+                assertionFailure("Failed to save user generated phrase: \(error)")
+            }
+        }
+        vc.modalPresentationStyle = .fullScreen
+        present(vc, animated: true)
+    }
+
+    private func installEmptyStateIfNeeded() {
+        guard AppConfig.emptyStatesEnabled else { return }
+        guard collectionView.backgroundView == nil else { return }
+        collectionView.backgroundView = PhraseCollectionEmptyStateView(action: addPhrasePressed)
+    }
+
+    private func removeEmptyStateIfNeeded() {
+        collectionView.backgroundView = nil
     }
     
     override func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
